Improve coding
